requirements.txt
fastapi[standard]==0.115.5
google-genai>=1.0.0
ffmpeg-python>=0.2.0
yt-dlp>=2024.0.0
python-dotenv==1.0.0
pydantic>=2.0.0
uvicorn>=0.30.0
